Create butterfly model stars using python
In this question, we need to print the pattern like below

*                 * 
* *             * * 
* * *         * * * 
* * * *     * * * * 
* * * * * * * * * * 
* * * * * * * * * * 
* * * *     * * * * 
* * *         * * * 
* *             * * 
*                 * 
Understanding the problem

We need to print 2 * n lines.
Center two lines have a pattern i.e print 2 * n asterisks( * ) respectively.
We need to divide it into two parts. one for the Upper part and the lower part.
A star is always followed by a space character.