Create Max and Min Value in List of Tuples             Write a program to print the maximum and minimum of all the values at index zero and index one in the list of tuples.                       Input  The first line of input will contain a positive integer (N). The next N lines will contain space-separated two integers, denoting elements of each tuple.                         Output  The first line of output should contain a tuple with the maximum and minimum value at index zero.  The second line of output should contain a tuple with the maximum and minimum value at index one. 
Explanation

For example, if the given N is 3, read the inputs in the next three
lines. If the inputs in the next three lines are the following.

1 5
3 2
5 8

As 5 is the maximum and 1 is the minimum in the values at index zero
and 8 is the maximum and 2 is the minimum in the values at index
one. So the output should be

(5, 1)
(8, 2)