Create Character Frequency                Write a program to compute the frequency of characters other than space.                         Input  The input will be a single line containing a string.  Note: 1. The input string will contain only alphabets and white space. 2. Ignore the case-sensitivity.                            Output  The output should be M lines equal to unique characters in the given string sorted in alphabetical order.  Each line should contain the character and its frequency separated by acolon (:). 
Explanation

For example, if the given string is "Pop up", all the unique characters
in the given string are (p, o, u). The characters (p, o, u) are with
frequency (3, 1, 1) respectively. Your code should print each unique
character and its frequency separated by a colon(:) in alphabetical
order. So the output should be

o: 1
p: 3
u: 1