Create Row and Column with Maximum                   For this problem, the prefilled code will contain an MxN matrix. Write a program to print the row and column containing the maximum element.                            Input  The first line of input will contain two space-separated integers, denoting the M and N.  The next M following lines will contain N space-separated integers, denoting the elements of each list.                            Output  The first line of output should be a list containing the row elements which has the maximum element.  The second line of output should be a list containing the column elements which has the maximum element. 
Explanation

For example, if the given M is 3 and N is 3, read the inputs in the next
three lines. If the numbers given in the next three lines are the
following.

1 2 3
10 20 30
5 10 15

As 30 is the maximum element in the given matrix, the first line of
output should be the second row and the second line of output
should be the third column. So the output should be

[10, 20, 30]
[3, 30, 15]